+++ app.py
@@ -1,5 +1,5 @@
-from flask import Flask, request, redirect, url_for, render_template, jsonify # importing flask and some of its kids because our whole app functions on them
-import string 
+from flask import Flask, request, redirect, url_for, render_template, jsonify  # importing flask and some of its kids because our whole app functions on them
+import string
 import random
 import sqlite3
 import os
@@ -7,23 +7,28 @@ from functools import wraps
 from urllib.parse import urlparse # url parsing library
 import hashlib # password hashing
 import datetime # date time stuff
+from flask_limiter import Limiter
+from flask_limiter.util import get_remote_address
 
-app = Flask(__name__) # inits flask
+app = Flask(__name__)
 
 DATABASE = 'url_shortener.db' # main database, stores URLs, access tokens and click counts
 BLOCKED_DOMAINS_DB = 'blocked_domains.db' # blocked domains database
 
-def get_db_connection(db_name): # establishes connection with two databases
-    conn = sqlite3.connect(db_name) 
+limiter = Limiter(
+    get_remote_address,
+    app=app,
+    default_limits=["200 per day", "50 per hour"]
+)
+
+def get_db_connection(db_name):
+    conn = sqlite3.connect(db_name)
     conn.row_factory = sqlite3.Row
     return conn
 
-def init_db(): # initializes both the database
+def init_db():
     with app.app_context():
         conn = get_db_connection(DATABASE)
-        # creates 2 tables in the url_shortener.db if it's not already there
-        # one table consists the basic stuff like ipa, original url, clicks, tokens
-        # the other stores tokens and information regarding them
         conn.execute('''
             CREATE TABLE IF NOT EXISTS url_mapping (
                 short_code TEXT PRIMARY KEY,
@@ -35,7 +40,7 @@ def init_db(): # initializes both the database
                 password TEXT
             )
         ''')
-        
+
         conn.execute('''
             CREATE TABLE IF NOT EXISTS api_tokens (
                 token TEXT PRIMARY KEY,
@@ -43,9 +48,8 @@ def init_db(): # initializes both the database
                 click_count INTEGER DEFAULT 0
             )
         ''')
-        
+
         conn.commit()
-        
         conn.close()
 
         conn = get_db_connection(BLOCKED_DOMAINS_DB) # connects to the blocked domains database
@@ -54,16 +58,11 @@ def init_db(): # initializes both the database
                 domain TEXT PRIMARY KEY
             )
         ''') # creates blocked domain database if it doesnt already exist
-        
-        conn.commit() 
-        
+
+        conn.commit()
         conn.close()
 
-
-# helper functions
-
-
-def is_valid_url(url): #checks if a url is valid using url parsing
+def is_valid_url(url):
     try:
         result = urlparse(url)
         return all([result.scheme, result.netloc])
@@ -74,13 +73,11 @@ def is_blocked_domain(url):
     conn = get_db_connection(BLOCKED_DOMAINS_DB)
     domain = urlparse(url).netloc
 
-    # Check if the domain or any of its subdomains are blocked
     blocked = conn.execute('SELECT 1 FROM blocked_domains WHERE ? LIKE (domain || \'.%\') OR domain = ?', (domain, domain)).fetchone()
     conn.close()
     return blocked is not None
 
 def generate_short_code(length, allow_numbers, allow_special, allow_uppercase, allow_lowercase):
-    # generates short code based on user's input
     characters = ''
 
     if allow_numbers:
@@ -101,65 +98,65 @@ def generate_short_code(length, allow_numbers, allow_special, allow_uppercase, a
     return ''.join(random.choice(characters) for _ in range(length))
 
 def require_api_token(f):
-    # Decorator function that restricts access to specific routes requiring a valid API token present in the Authorization header.
     @wraps(f)
     def decorated_function(*args, **kwargs):
         token = request.headers.get('Authorization')
         if not token:
-            return 'API token is missing', 401 # returns a 401 if token is missing
+            return 'API token is missing', 401
 
         conn = get_db_connection(DATABASE)
         valid_token = conn.execute('SELECT 1 FROM api_tokens WHERE token = ?', (token,)).fetchone()
-        # would only let go if the token is in the database
         conn.close()
 
         if not valid_token:
-            return 'Invalid API token', 403 # returns a 403 if token is not valid
+            return 'Invalid API token', 403
 
         return f(*args, **kwargs)
     return decorated_function
 
-@app.route('/') #main route, home page
+@app.route('/')
 def home():
     return render_template('index.html')
 
-@app.route('/gen') # route to the short URL generator page
-def index(): 
-    conn = get_db_connection(DATABASE) # connects with the database
-    urls = conn.execute('SELECT * FROM url_mapping').fetchall() # fetches all URLs from the database
+@app.route('/gen')
+def index():
+    conn = get_db_connection(DATABASE)
+    urls = conn.execute('SELECT * FROM url_mapping').fetchall()
     conn.close()
 
-    return render_template('gen.html', urls=urls) # shows the database in the html file
+    return render_template('gen.html', urls=urls)
 
-@app.route('/shorten', methods=['POST']) # handles form submissions and generates URLs
+@app.route('/api')
+def api():
+    return render_template('api.html')
+
+@app.route('/shorten', methods=['POST'])
+@limiter.limit("3 per 5 second")  # Rate limit for /shorten route
 def shorten():
     original_url = request.form['url']
     custom_code = request.form.get('custom_code')
     password = request.form.get('password')
     expiry_time = request.form.get('expiry_time')
 
+    if 'Authorization' in request.headers:
+        return 'This route is not intended for API use', 403
+
     length = int(request.form['length'])
     allow_numbers = 'allow_numbers' in request.form
     allow_special = 'allow_special' in request.form
     allow_uppercase = 'allow_uppercase' in request.form
     allow_lowercase = 'allow_lowercase' in request.form
-# extracts all details from the form
 
     if not is_valid_url(original_url):
         return 'Invalid URL', 400
-    # if an invalid URL was given, a 400 is shown
 
     if is_blocked_domain(original_url):
         return 'Blocked domain', 400
-    # if a domain is blocked, a 400 is shown
 
     ip_address = request.remote_addr
-    # fetches the IP address
 
     conn = get_db_connection(DATABASE)
-    # gets connection to the database
 
-    # starts short code generation
     characters = ''
     if allow_numbers:
         characters += string.digits
@@ -170,69 +167,69 @@ def shorten():
     if allow_lowercase:
         characters += string.ascii_lowercase
 
-    if not characters: # if no custom input was provided
-        characters = string.ascii_letters + string.digits #big letter small letter and numbers
+    if not characters:
+        characters = string.ascii_letters + string.digits
 
-    if not expiry_time: # if no expiry time was provided
-        expiry_time = None # new variable
-        
-    total_combinations = len(characters) ** length # checks for total combination possible
+    if not expiry_time:
+        expiry_time = None
 
-    existing_codes_count = conn.execute('SELECT COUNT(*) FROM url_mapping').fetchone()[0] # counts total combinationation available
-    
-    if existing_codes_count >= total_combinations: # if no other combinations are possible
+    total_combinations = len(characters) ** length
+
+    existing_codes_count = conn.execute('SELECT COUNT(*) FROM url_mapping').fetchone()[0]
+
+    if existing_codes_count >= total_combinations:
         options_message = f"Length: {length}, Allow Numbers: {allow_numbers}, Allow Special: {allow_special}, Allow Uppercase: {allow_uppercase}, Allow Lowercase: {allow_lowercase}"
         return f'No other possible combination of {length} characters with the selected options is available. Selected options: {options_message}', 400
 
-    if custom_code: # if a custom code was provided, checks if it was already given
+    if custom_code:
         short_code = custom_code
         if conn.execute('SELECT 1 FROM url_mapping WHERE short_code = ?', (short_code,)).fetchone() is not None:
             return 'Custom code already exists. Please choose another one.', 400
-    else: # or else generates a short code
+    else:
         short_code = generate_short_code(length, allow_numbers, allow_special, allow_uppercase, allow_lowercase)
         while conn.execute('SELECT 1 FROM url_mapping WHERE short_code = ?', (short_code,)).fetchone() is not None:
             short_code = generate_short_code(length, allow_numbers, allow_special, allow_uppercase, allow_lowercase)
 
-    if password: # if a password was given, hashes it
+    if password:
         password_hash = hashlib.sha256(password.encode()).hexdigest()
-    else: 
-        password_hash = None #if not, then does nothing
+    else:
+        password_hash = None
 
     conn.execute('INSERT INTO url_mapping (short_code, original_url, ip_address, expiry_time, password) VALUES (?, ?, ?, ?, ?)',
-                 (short_code, original_url, ip_address, expiry_time, password_hash)) # commits to the database with required information
-    conn.commit() 
-    conn.close() # closes connection
+                 (short_code, original_url, ip_address, expiry_time, password_hash))
+    conn.commit()
+    conn.close()
 
     return redirect(url_for('index'))
 
 @app.route('/<code>')
-def redirect_to_url(code): # for the /shortened_url redirection part
+def redirect_to_url(code):
     conn = get_db_connection(DATABASE)
-    url = conn.execute('SELECT original_url, click_count, expiry_time, password FROM url_mapping WHERE short_code = ?', (code,)).fetchone() # checks which url is the short code linked to
+    url = conn.execute('SELECT original_url, click_count, expiry_time, password FROM url_mapping WHERE short_code = ?', (code,)).fetchone()
 
-    if url: # if url exists
-        if url['expiry_time'] and datetime.datetime.now() > datetime.datetime.strptime(url['expiry_time'], '%Y-%m-%d %H:%M:%S'): # but it is expired
-            conn.execute('DELETE FROM url_mapping WHERE short_code = ?', (code,)) # delete the URL mapping
-            conn.execute('UPDATE api_tokens SET link_count = link_count - 1 WHERE token = (SELECT api_token FROM url_mapping WHERE short_code = ?)', (code,)) # delete the links in both API and stuff
-            conn.commit() 
+    if url:
+        if url['expiry_time'] and datetime.datetime.now() > datetime.datetime.strptime(url['expiry_time'], '%Y-%m-%d %H:%M:%S'):
+            conn.execute('DELETE FROM url_mapping WHERE short_code = ?', (code,))
+            conn.execute('UPDATE api_tokens SET link_count = link_count - 1 WHERE token = (SELECT api_token FROM url_mapping WHERE short_code = ?)', (code,))
+            conn.commit()
             conn.close()
             return 'URL has expired', 404
 
-        if url['password']: # and url has password
-            return render_template('password.html', code=code) # redirects to the password.html
+        if url['password']:
+            return render_template('password.html', code=code)
 
-        conn.execute('UPDATE url_mapping SET click_count = click_count + 1 WHERE short_code = ?', (code,)) # increases the click count by 1
-        conn.execute('UPDATE api_tokens SET click_count = click_count + 1 WHERE token = (SELECT api_token FROM url_mapping WHERE short_code = ?)', (code,)) # in api too 
+        conn.execute('UPDATE url_mapping SET click_count = click_count + 1 WHERE short_code = ?', (code,))
+        conn.execute('UPDATE api_tokens SET click_count = click_count + 1 WHERE token = (SELECT api_token FROM url_mapping WHERE short_code = ?)', (code,))
         conn.commit()
         conn.close()
-        return redirect(url['original_url']) # redirects to the url
+        return redirect(url['original_url'])
 
-    else: # if url was not found
+    else:
         conn.close()
-        return 'URL not found', 404 # returns a 404
+        return 'URL not found', 404
 
-@app.route('/check_password', methods=['POST']) # checks for passwords
-def check_password(): 
+@app.route('/check_password', methods=['POST'])
+def check_password():
     code = request.form['code']
     password = request.form['password']
 
@@ -249,10 +246,10 @@ def check_password():
     conn.close()
     return 'Incorrect password', 403
 
-@app.route('/api/shorten', methods=['POST']) #API routes
-
-@require_api_token # requires the goddamn token
-def api_shorten(): #shortens url but using API
+@app.route('/api/shorten', methods=['POST'])
+@require_api_token
+@limiter.limit("50 per minute")  # Rate limit for /api/shorten route
+def api_shorten():
     data = request.json
     original_url = data.get('url')
     length = int(data.get('length', 6))
@@ -305,7 +302,7 @@ def api_shorten(): #shortens url but using API
 
     return {'short_code': short_code}, 201
 
-@app.route('/api/generate_token', methods=['POST']) # Generates a goddamn token for you
+@app.route('/api/generate_token', methods=['POST'])
 def generate_token():
     token = ''.join(random.choices(string.ascii_letters + string.digits, k=32))
